Cache types in atoms and expressions.
This lets us query types without an environment. The reason to do this now is
that we're about to allow decls on the rhs of pi types, which means that
evaluating the type of `f(x, y)` will require emitting decls. We don't want to
have to emit decls every time we query the type of an expression so we need to
have the type cached. And if we're doing that, we may as well go all the way and
require that the type be queryable even without an environment. That also lets
us take away the type annotation from decls, which reduces redundancy a bit. As
a bonus, we can now query the types of terms in i-space, so we don't need
`getTypeSubst` and similar.