Small tweak to the parse/group highlighting.
Highlighting every parent group in a slightly different shade was just too hard
to interpret. I changed it to only show the innermost non-leaf node containing
the pointed-at character (in light blue) along with its immediate children (in
yellow).

Also removed the span IDs because we're not using them and we need to figure out
the story around uniqueness.